Uploaded the old project
This is an old project I did back when I first was learning to program
years ago. Currently it's hardcoded to connect only to ucsc servers, but
with a few tweaks it can be made to connect to anywhere.
NOTE: This has not been touched in years, and doesn't reflect the vast
amount of knowledge I've learned since then. This just shows my
knowledge level after 6 months of Java